requirements-all.txt
-r requirements-apps-api.txt
-r requirements-apps-process-new-granules.txt
-r requirements-apps-scale-cluster.txt
-r requirements-apps-start-execution.txt
-r requirements-apps-update-db.txt

aws-cdk-lib==2.2.0
constructs>=10.0.0,<11.0.0

boto3==1.20.24
jinja2==3.0.3
moto==1.3.14
pytest==6.2.5
PyYAML==6.0
responses==0.16.0
flake8==4.0.1
flake8-import-order==0.18.1
flake8-blind-except==0.2.0
flake8-builtins==1.5.3
openapi-spec-validator==0.3.1
click==8.0.3
prance==0.21.8.0
cfn-lint==0.56.3
